pyproject.toml
[tool.poetry]
name = "target-bigquery"
version = "0.0.1"
description = "`target-bigquery` is a Singer target for BigQuery, built with the Meltano SDK for Singer Targets."
authors = ["Jonas Kalderstam"]
keywords = [
    "ELT",
    "BigQuery",
]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "<3.11,>=3.7.1"
singer-sdk = "^0.11.1"
google-cloud-bigquery = "~3.3.3"
fastavro = "^1.6.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
tox = "^3.24.4"
flake8 = "^3.9.2"
black = "^21.9b0"
pydocstyle = "^6.1.1"
mypy = "^0.910"
types-requests = "^2.26.1"
isort = "^5.10.1"

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "target_bigquery"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
target-bigquery = 'target_bigquery.target:TargetBigQuery.cli'
